add animetail

+++ data/src/main/java/tachiyomi/data/track/anime/AnimeTrackMapper.kt
@@ -2,7 +2,21 @@ package tachiyomi.data.track.anime
 
 import tachiyomi.domain.track.anime.model.AnimeTrack
 
-val animeTrackMapper: (Long, Long, Long, Long, Long?, String, Double, Long, Long, Float, String, Long, Long) -> AnimeTrack =
+val animeTrackMapper: (
+    Long,
+    Long,
+    Long,
+    Long,
+    Long?,
+    String,
+    Double,
+    Long,
+    Long,
+    Double,
+    String,
+    Long,
+    Long,
+) -> AnimeTrack =
     { id, animeId, syncId, remoteId, libraryId, title, lastEpisodeSeen, totalEpisodes, status, score, remoteUrl, startDate, finishDate ->
         AnimeTrack(
             id = id,
