+++ domain/src/main/java/tachiyomi/domain/track/anime/model/AnimeTrack.kt
@@ -10,7 +10,7 @@ data class AnimeTrack(
     val lastEpisodeSeen: Double,
     val totalEpisodes: Long,
     val status: Long,
-    val score: Float,
+    val score: Double,
     val remoteUrl: String,
     val startDate: Long,
     val finishDate: Long,
